Added first working version of script for graphenic sheet
I added the first working script that drives the simulation of a sample with only graphenic sheets (GSs), which in the C++ code are referred to as graphene nanoplatelets (GNPs).

This version only works with GS that lie entirely inside the sample. This beacuase the functions that perform the operations of cutting the parts outside the sample are commented.