Added functions for CNTs to script for mixed fillers
Copied all the functions realted to CNTs that were missing in the script for mixed fillers. I added functions that were related mainly with the generation of the model in Abaqus (Abaqus functions), and also string and math functions. String functions are used to generated names as strings. Math functions, as its name indicates, perform some mathematical operations.