Corrected script for mixed fillers and for GSs
Made some corrections to the scrip for mixed fillers after I tried to run some simulations in Aabqus. Some of these changes were also relevant to the script for GSs, thus the same corrections were added to that script.

I made an additional change for dealing with the case when the "minimum point" of the sample is not the origin. For this I added a translation when generating the instance of the matrix. This caused the function that generated sets for the periodic boundary conditions (PBCs) to cause errors. This was because the faces of the RVE were not found due to a misinterpretation of the meaning of variable Lxyz. This only happened when the minimum point of the sample was not the origin. Thus, the function that generated the sets for the PBCs was corrected.

I added more input parameters, which are actually known from the input file. These are the maximum dimensions of the GSs. Previously a GS was used, however this could cause errors in the future if a distribution of dimensions is added.