Corrected generation of BiggerBox and meshing
Corrected the way BiggerBox was created. This was done assuming P0 of the RVE was always the origin. So, now it is constructed taking as reference its own P0. I did this change to the scripts for only GNPs and mixed fillers.

I made other changes in the script for only GNPs:

1) Added a Zero variable

2) Meshing of the matrix was being done inside a loop. Of course, this only needs to be done once. So now the meshing of the matrix is done outside the loop.

3) In the definition of function Create_Set_for_PBC an argument was missing. Thus, it was added.