Corrected equations for PBCs
The equations to set periodic boundary conditions (PBCs) needed two corrections. These corrections were made only to the script for mixed particles and only GSs since the script for only CNTs does not have equations for PBCs (for now).

First, a negative sign was needed on the reference point because without it a positive displacement resulted in compression. Now, with the negative sign applied on the reference point a positive displacement results in tensions.

Second, the degrees of freedom of reference points for the y- and z-directions were incorrectly assigned. They were assign to direction 1 (i.e., x). Now they are assigned to directions 2 and 3, respectively.