Corrected condition for remeshing GSs
One of the two conditions needed for a GS to be remeshed was not properly implemented. The condition checked the number of meshed regions in the assembly but always was zero. Now the condition checks the number of nodes in the part and it is zero when the GS was not meshed at first. This is how the condition should work.

This update was implemented on the scripts for GSs and for mixed particles.