Added extended RVE for CNT networks
Some parts of the embedded elements for th eCNTs lie outside the RVE. This causes Abaqus to throw an error that there are embedded elements without host element. In order to avoid this error, I extended the RVE to ensure that even if there are CNTs outside the RVE they still have a host element.