Added extended RVE to script for mixed fillers
Added functions and variables needed to include the extended RVE for the case of mixed fillers. The extended RVE prevents an error from happening when parts of CNT embedded elments lie outside the RVE.

I also modified some functions since now I need to create the extended RVE and the RVE. The extended RVE is used in the simulation, but the RVE is used to create the cutter object (tu trim GNPs).