Missing hide set and undefined variables in mixed fillers script
I had not implemented the function to create the set that contains all elments in the extended layer of the RVE. Of course this is needed to hide them in the visualization module. Now this function was added. I also changed the variable for the set name from global to being local to the function that creates this set. Set names are not global so this change keeps consistency of the script.

Also some variables were not defined or had a name used in other scripts. So now all variables are consistent.