Changed variable names in scrip for CNTs
I changed the variable str_matrix to matrixName in the main program in order to make it consistent with the naming in the scripts for GNPs and mixed particles.

I also changed the scope of the variable for the set that contains the elements in the extended layer of the RVE. Now this variable is local to the function that creates this set instead of being global. Variables for set names are not global so this change keeps consistency of the script.