Homogenized message of number of particles in the simulation
So far, only the script for GNPs and mixed fillers printed a message with the numbe of GNPs in the simulation. So I added the number of CNTs for the mixed fillers script and also for the script for CNTs. In this way the massage with numbe of particles in homogenized among all three scripts.

In addition I changed the name of the number of GNPs in the script for only GNPs so that it matches the name in the script for mixef fillers.

Another small change is that there was a ')' that I forgot to delete in the previous update.