Added time measurements for parts of the script
I added time measurements for centain parts of the script. I tried to be consistent however there are some differences on how GNPs and CNTs are created that does not allow for measuring time in at the same check points in all three scripts.

I also added the possibility to measure the time execution of the job and the whole Abaqus simulation.

In the case of CNTs I changed the name of the job to be more consistent with the other scripts.