Corrected creation of sets for embedded elements
The sets used to add the embedded elements constrains should contain the whole CNT. Some CNTs are cut to facilitate meshing. This caused some CNTs to consist of more than two cells. When creating the sets for embedded elements only one cell was considered. This caused some cells to not have the embedded elments constraint and result in a error where elments were missing properties.

Now all cells are considered when creating the sets for embedded elements constraints.