Changed how sets for CNT nodes are generated
I changed how the sets for CNT nodes are generated. Before I was finding nodes and merging sets for each new node that was found. Now, each node that is found is appended to an array. Then the set is made from this array of nodes. This new way to generate the set is much faster. I guess the merge operation from Abaqus is too slow copared with the append method from Python.