Added function for printing messages to terminal and file
Created a function to send messages to three sources: the Abaqus message window, the terminal, and to a file. I created this functions because when running a simulation from the terminal, the messages using the "print" function from Python are lost. With the new function, independently of how the simulation is run the messages are always visible in the interface used and in a file.