Changed some names to upper case letters only
For the modified script where I add parts to a model imported from an Abaqus input file, I need to use set names that are all in upper case letters. Thus I modified some names of sets, materials and sections to be in all caps in order to avoid errors.