Handling of a CNTs that cannot be swept
Sometimes a CNT cannot be sewpt and Abaqus trhows an error. So far, the cases when this happens can be avoided by using a flag when performing the sweep: profileNormal=ON. Thus, I added a try-except case to catch the Abaqus error and try the swept again with the necessary flag.

This change affects the scripts for CNTs and mixed fillers. In the case of CNTs I had to update the indentation in the function that performs the sweep.