Finding and remeshing GSs with zero volume elements
Sometimes there are errors of zero volume elements on GSs that are cut. This is because after cutting a GS it cannot have a regular mesh. In such cases the GS is meshed using tetrahedrons. So I added a function to find these GSs and remeshing them using tetrahedrons. This is done after performing a data check.