Added variations of scripts that handle issues after a simulation is done
I created some scripts that handle specific issues after a simulation is done. That is, these script create a new model from an existing input file, then fix the issues and save a new input file with the issue fixed.

The code for fixing such issues cannot be in the main scripts since those issues can be (and are) solved during generation. If the simulation is done and we only have and input file with the issues present, another strategy is needed, i.e., these scripts.

In this update two scripts are added:

a) Abaqus_redo_cnt_set.py: Sometimes Abaqus adds two nodes in place of a CNT point. This creates a set that has more nodes than CNT points. This in turn causes an error in the C++ code since it is expecting the same number of nodes as CNT points. Thus, this script finds the sets that have more nodes than CNT points and then removes the extra nodes.

b) Abaqus_remesh_gss.py: Some GSs that are cut, when meshed, have elements that result in the error of elements with zero volume. This scripts finds those GSs and remeshes them with tetrahedrons to eleminate such an error. For this to work, an odb file with the set that contains elements with such an error is needed.