Additional cuts on CNT cells
Sometimes CNT cells that need to be cut to facilitate meshing cannot be cut. Those points are saved to make a second attempt to cut the CNT cell. This is done in this update by saving the point that resulted in an invalid cut and the previous and next points. Then cuts are made on all the saved points. Some repetitions are discarded.