Updated data check functionality for CNTs
I updated the way the data check works for CNTs so that it matches the other two scripts. The flag for data check only worked when the flag for saving the input file was active. Now the flag for data check is independent from the value of the flag for saving the input file.

I also remove some comments that did not represent how the data check flag worked.