Update on functions for CNT sets
In the function that creates the sets for CNT points I forgot to save a set into a variable. This caused an error since a variable was being used without being defined first. So I corrected this.

I also added some line breaks in some strings used in the plog function within the functions that creates the CNT points.