Corrected calculation of vector to find an octagon
To cut a CNT cell, a set of edges forming an octagon are needed. To find those edges, some calculations are made to determine a cylinder that encloses such edges. In the process of determining the cylinder, some vectors are calculated and sometimes they need to be unit vectors. There were two vectors that were not converted to unit vectors but needed to be. Thus, I make those two vectors unitary.