Update on how to delete embedded element constraints
When the input file is generated by parts and there are GSs, the constraints are not numbered according to the numbering of CNTs. Thus, when deleting constraints the data check resulted in errors regarding the embedded element constraints. To correct this I created a dictionary that maps an instance with its constraint.