Correcting the case of empty CNT sets
Sometimes, a CNT is meshed but with an irregular mesh. This causes to not have nodes in place of the CNT points. It seems this happens because some segments are longer than the input segment length, so Abaqus places more elements in those longer segments. These longer segments in turn come from those CNTs at boundaries when a short CNT segment (cut by the sample boundary) is merged with a regular CNT segment.

I found this issue is solved by cutting the first or last CNT segment. This hepls Abaqus create a regular mesh. So I added some functions to find empty sets and then cut the CNT cell on the second and penultimate points (i.e., first and last segments). Then, the CNT is meshed again and the set with the CNT nodes is created again.

These changes were made to the scripts for CNTs and mixed fillers.