Updated CNT cell partitioning for empty set
I found that when a CNT was meshed with an irregular mesh so that the set with CNT nodes is empty, some times it is not enough to make a single cut to mesh the CNT. Sometimes a segment of the CNT is not meshed. This results in a part consisting of two CNT segments which causes errors in the C++ code. Thus, it is best to cut the cell on all points to ensure a proper mesh. Here I updated the scripts for CNTs and mixed fillers. I also updated the script that creates CNTs again.